Reference

Reference

Contents

Reference

Core

Nonlinear

MLP

RNN

Function Index

Core

KLayers.Embed — Type.

Embed(inputSize,embedSize;winit=xavier)

Creates and embedding layer according to given inputSize and embedSize.

By default embedding parameters initialized with xavier, you can change it winit argument

(m::Embed)(x::Array{T}) where T<:Integer
(m::Embed)(x)

Embed objects are callable with an input which is either and integer array (one hot encoding) or an N-dimensional matrix. For N-dimensional matrix, size(x,1)==inputSize

source

KLayers.Linear — Type.

Linear(inputSize,outputSize;kwargs...)
(m::Linear)(x) #forward run

Creates and linear layer according to given inputSize and outputSize.

By default embedding parameters initialized with xavier, you can change it winit argument

Keywords

winit=xavier: weight initialization distribution

bias=zeros: bias initialization distribution

source

KLayers.Dense — Type.

Dense(inputSize,outputSize;kwargs...)
(m::Dense)(x) #forward run

Creates and deense layer according to given inputSize and outputSize.

Keywords

winit=xavier: weight initialization distribution

bias=zeros: bias initialization distribution

f=ReLU(): activation function

source

KLayers.Conv — Type.

Conv(h,[w,c,o];kwargs...)
(m::Conv)(x) #forward run

Creates and convolutional layer according to given filter dimensions.

Keywords

winit=xavier: weight initialization distribution

bias=zeros: bias initialization distribution

padding=0: the number of extra zeros implicitly concatenated at the start and at the end of each dimension.

stride=1: the number of elements to slide to reach the next filtering window.

upscale=1: upscale factor for each dimension.

mode=0: 0 for convolution and 1 for cross-correlation.

alpha=1: can be used to scale the result.

handle: handle to a previously created cuDNN context. Defaults to a Knet allocated handle.

source

KLayers.BatchNorm — Type.

BatchNorm(channels:Int;o...)
(m::BatchNorm)(x;o...) #forward run

source

Nonlinear

KLayers.ReLU — Type.

ReLU()
(l::ReLU)(x) = max(0,x)

Fast kernel is avaiable for gpu

source

KLayers.Sigm — Type.

Sigm()
(l::Sigm)(x) = sigm(x)

Sigmoid function

source

KLayers.Tanh — Type.

Tanh()
(l::Tanh)(x) = tanh(x)

Tangent hyperbolic function

source

KLayers.ELU — Type.

ELU()
(l::ELU)(x) = elu(x) -> Computes x < 0 ? exp(x) - 1 : x

source

KLayers.LeakyReLU — Type.

LeakyReLU(α=0.2)
(l::ELU)(x) -> Computes x < 0 ? α*x : x

source

KLayers.Dropout — Type.

Dropout(p=0)

Dropout Layer. p is the droput probability.

source

KLayers.SoftMax — Type.

SoftMax(dims=:)
(l::SoftMax)(x)

Treat entries in x as as unnormalized scores and return softmax probabilities.

dims is an optional argument, if not specified the normalization is over the whole x, otherwise the normalization is performed over the given dimensions. In particular, if x is a matrix, dims=1 normalizes columns of x and dims=2 normalizes rows of x.

source

KLayers.LogP — Type.

LogP(dims=:)
(l::LogP)(x)

Treat entries in x as as unnormalized log probabilities and return normalized log probabilities.

dims is an optional argument, if not specified the normalization is over the whole x, otherwise the normalization is performed over the given dimensions. In particular, if x is a matrix, dims=1 normalizes columns of x and dims=2 normalizes rows of x.

source

KLayers.LogSumExp — Type.

LogSumExp(dims=:)
(l::LogSumExp)(x)

Compute log(sum(exp(x);dims)) in a numerically stable manner.

dims is an optional argument, if not specified the summation is over the whole x, otherwise the summation is performed over the given dimensions. In particular if x is a matrix, dims=1 sums columns of x and dims=2 sums rows of x.

source

MLP

KLayers.MLP — Type.

MLP(h::Int...;kwargs...)

Creates a multi layer perceptron according to given hidden states. First hidden state is equal to input size and the last one equal to output size.

(m::MLP)(x;prob=0)

Runs MLP with given input x. prob is the dropout probability.

Keywords

winit=xavier: weight initialization distribution

bias=zeros: bias initialization distribution

f=ReLU(): activation function

source

RNN

KLayers.RNN — Type.

SRNN(inputSize,hiddenSize;activation=:relu,options...)
LSTM(inputSize,hiddenSize;options...)
GRU(inputSize,hiddenSize;options...)

(1) (l::T)(x;kwargs...) where T<:RNN
(2) (l::T)(x::Array{Int};batchSizes=nothing,kwargs...) where T<:RNN
(3) (l::T)(x::Vector{Vector{Int}};sorted=false,kwargs...) where T<:RNN

All RNN layers has above forward run(1,2,3) functionalities.

(1) x is an input array with size equals d,[B,T]

(2) For this You need to have an RNN with embedding layer. x is an integer array and inputs coressponds one hot vector indices. You can give 2D array for minibatching as rows corresponds to one instance. You can give 1D array with minibatching by specifying batch batchSizes argument. Checkout Knet.rnnforw for this.

(3) For this You need to have an RNN with embedding layer. x is an vector of integer vectors. Every integer vector corresponds to an instance. It automatically batches inputs. It is better to give inputs as sorted. If your inputs sorted you can make sorted argument true to increase performance.

Outputs of the forward functions are always y,h,c,indices. h,c and indices may be nothing depending on the kwargs you used in forward.

y is last hidden states of each layer. size(y)=(H/2H,[B,T]). If you use batchSizes argument y becomes 2D array size(y)=(H/2H,sum(batchSizes)). To get correct hidden states for an instance in your batch you should use indices output.

h is the hidden states in each timesstep. size(h) = h,B,L/2L

c is the hidden states in each timesstep. size(h) = h,B,L/2L

indices is corresponding indices for your batches in y if you used batchSizes. To get ith instance's hidden states in each times step, you may type: y[:,indices[i]] `

options

embed=nothing: embedding size or and embedding layer

numLayers=1: Number of RNN layers.

bidirectional=false: Create a bidirectional RNN if true.

dropout=0: Dropout probability. Ignored if numLayers==1.

skipInput=false: Do not multiply the input with a matrix if true.

dataType=Float32: Data type to use for weights.

algo=0: Algorithm to use, see CUDNN docs for details.

seed=0: Random number seed for dropout. Uses time() if 0.

winit=xavier: Weight initialization method for matrices.

binit=zeros: Weight initialization method for bias vectors.

usegpu=(gpu()>=0): GPU used by default if one exists.

kwargs

hx=nothing : initial hidden states

cx=nothing : initial memory cells

hy=false : if true returns h

cy=false : if true returns c

source

Function Index

KLayers.BatchNorm

KLayers.Conv

KLayers.Dense

KLayers.Dropout

KLayers.ELU

KLayers.Embed

KLayers.LeakyReLU

KLayers.Linear

KLayers.LogP

KLayers.LogSumExp

KLayers.MLP

KLayers.RNN

KLayers.ReLU

KLayers.Sigm

KLayers.SoftMax

KLayers.Tanh